Shift & transform support
As suspected, we need to adjust the kernel coordinate system to make this
work. It is easier to implement it this way than try to correct the kernel
afterwards (especially for the image transformation!). Also I *think* this
is going to make it possible to combine this whole thing correctly with
anti-aliasing kernels.