Imaging-dask changes
Reworked visibility graphing. It had been iterating at a too low level. I changed the structure and can now do larger numbers of visibility sets.

Added some asserts to check on types on entry of key functions. One failure mode of working with graphs is to pass a list of items instead of the item itself.