Facets and wprojection now working
Had sense of calculation of the delay centre reversed.

Also had to untangle some circular imports for arl.visibility.operations

For facets and timeslice, the imaging test results are so poor that it’s difficult to test properly. For facets, the tests with wstack and wprojection also test faceting. So I’ve all the facets and timeslice tests do is execute and do minimal checks.

At the testing scale some of these are very inefficient. Some simple changes are need to maintain context across function calls. For example, the wstack/wprojection combo recalculates the w kernel for each wstack.