Pipeline timings
New script for standard testing. In examples/performance.

This runs through predict, invert, and ical. The scale is adapted to the number of processors available and the memory. These heuristics may need overriding in special cases.

A variety of cases are run and the results output to a csfile.

Currently only works for single machine.

Expect a single run to take a few hours.

Note that the requirements have been updated to include seqfile.