Substantial changes to make various forms of distribution consistent
Now have functions, iterations, bags, and delayed, as ways of iterating in calibration and imaging.

Also includes ICAL pipeline for SIP usage. Found in a notebook imaging-pipelines-sip and also in testing-pipelines-functions.

Also worked to reduce the number of various of similar test programs.

Still needs more work to control some entropy but this is a start.