Calibration context introduced
Specification of options for calibration (as in ICAL) was getting to be complicated. I introduced calibration context to hold the options for calibration. Thus T, G, P, B, and I calibrations can be specified and controlled.

Also timeslice iterator for gaintable.

Fixed solve_gaintable failure of logic when solving for intervals longer than the gap between integrations. Solving for B triggered this.