skymodel_cal can now work with images
skymodel_cal can now work with images

This means that one of the SkyModels can be an image, and the gains solved for. Next step is to solve for the images.

Only the demonstration notebook changed.

Fixed predict not converting vis to blockvis on output (triggered if input was blockvis).