Explicitly indicate that sequence is expected
In general, when a sequence is expected one should use nout=N when
calling daks's (and daliuge's) delayed. Dask happily works *without* the
nout specification, but daliuge currently doesn't. If this situation
starts happening more ubiquitously we could spend a bit more of effort
and support it in daliuge as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>